feature: avatar independed component and some profile refacroring
