+++ src/pages/Profile/Avatar/Avatar.module.scss
@@ -1,0 +1,68 @@
+@import "src/styles/partials/theme";
+
+.avatar {
+  height: 300px;
+  width: 300px;
+  border-radius: 50%;
+  border: 4px solid $warning-200;
+  text-align: center;
+  background-color: $background-dark-10;
+  position: relative;
+  transition: background-color $transition-delay;
+  flex-shrink: 0;
+
+  input[type="file"] {
+    display: none;
+  }
+
+  img {
+    width: 100%;
+    height: 100%;
+    border-radius: 50%;
+    object-fit: cover;
+  }
+
+  .addPhoto {
+    position: absolute;
+    top: 0;
+    left: 0;
+    right: 0;
+    bottom: 0;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    background: $black;
+    border-radius: 50%;
+    opacity: 0;
+    transition: 0.5s;
+    cursor: pointer;
+
+    .camera {
+      font-size: 80px;
+      color: $gray-600;
+    }
+
+    &:hover {
+      opacity: 0.5;
+    }
+  }
+}
+
+:global(.light_theme) .avatar {
+  background-color: $gray-500;
+}
+
+@media screen and (max-width: 1480px) {
+  .avatar {
+    width: 280px;
+    height: 280px;
+  }
+}
+
+@media screen and (max-width: 768px) {
+  .avatar {
+    margin-top: 20px;
+    width: 160px;
+    height: 160px;
+  }
+}
