+++ src/pages/Profile/Profile.module.scss
@@ -63,73 +63,6 @@
   background-color: $gray-600;
 }
 
-.idWrapper {
-  position: absolute;
-  top: 15px;
-  right: 15px;
-  font-weight: 600;
-  cursor: pointer;
-  color: $white;
-  transition: background-color $transition-delay, color $transition-delay;
-
-  &:hover {
-    color: $warning-200;
-  }
-}
-
-:global(.light_theme) .idWrapper {
-  color: $black;
-}
-
-.buttons {
-  position: absolute;
-  bottom: 0;
-  width: 100%;
-  left: 0;
-  right: 0;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  padding: 7px;
-
-  .exit {
-    font-weight: 500;
-    white-space: nowrap;
-    cursor: pointer;
-    color: $white;
-    transition: background-color $transition-delay;
-    @include style-button-icon-color($gray-600,
-      $error-200,
-      $error-300,
-      $error-200 );
-  }
-
-  .button {
-    color: $gray-600;
-    font-weight: 500;
-    white-space: nowrap;
-    cursor: pointer;
-    color: $white;
-    transition: background-color $transition-delay;
-
-    &:hover {
-      color: $warning-200;
-    }
-  }
-
-  .delete {
-    font-weight: 500;
-    white-space: nowrap;
-    cursor: pointer;
-    color: $white;
-    transition: background-color $transition-delay;
-    @include style-button-icon-color($gray-600,
-      $error-200,
-      $error-300,
-      $error-200 );
-  }
-}
-
 .avatar {
   height: 300px;
   width: 300px;
@@ -180,6 +113,91 @@
 
 :global(.light_theme) .avatar {
   background-color: $gray-500;
+}
+
+.idWrapper {
+  position: absolute;
+  top: 15px;
+  right: 15px;
+  font-weight: 600;
+  cursor: pointer;
+  color: $white;
+  transition: background-color $transition-delay, color $transition-delay;
+
+  &:hover {
+    color: $warning-200;
+  }
+}
+
+:global(.light_theme) .idWrapper {
+  color: $black;
+}
+
+.buttons {
+  position: absolute;
+  bottom: 0;
+  width: 100%;
+  left: 0;
+  right: 0;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  padding: 7px;
+
+  .exit {
+    font-weight: 500;
+    white-space: nowrap;
+    cursor: pointer;
+    color: $white;
+    transition: background-color $transition-delay;
+    @include style-button-icon-color(
+      $gray-600,
+      $error-200,
+      $error-300,
+      $error-200
+    );
+  }
+
+  .button {
+    color: $gray-600;
+    font-weight: 500;
+    white-space: nowrap;
+    cursor: pointer;
+    color: $white;
+    transition: background-color $transition-delay;
+
+    &:hover {
+      color: $warning-200;
+    }
+  }
+
+  .delete {
+    font-weight: 500;
+    white-space: nowrap;
+    cursor: pointer;
+    color: $white;
+    transition: background-color $transition-delay;
+    @include style-button-icon-color(
+      $gray-600,
+      $error-200,
+      $error-300,
+      $error-200
+    );
+  }
+}
+
+:global(.light_theme) .buttons {
+  .exit {
+    color: $black;
+  }
+
+  .button {
+    color: $black;
+  }
+
+  .delete {
+    color: $black;
+  }
 }
 
 .info {
@@ -320,20 +338,6 @@
   }
 }
 
-:global(.light_theme) .buttons {
-  .exit {
-    color: $black;
-  }
-
-  .button {
-    color: $black;
-  }
-
-  .delete {
-    color: $black;
-  }
-}
-
 .passwordWrapper {
   display: flex;
   flex-direction: row;
@@ -381,7 +385,6 @@
 
         .info {
           .line {
-
             .text,
             .name,
             .inputName {
@@ -393,7 +396,6 @@
     }
   }
 }
-
 
 @media screen and (max-width: 768px) {
   .wrapper {
@@ -417,13 +419,6 @@
       max-width: none;
 
       .row {
-        .buttons {
-          display: flex;
-          flex-direction: column;
-          position: relative;
-          gap: 30px;
-        }
-
         .info {
           .line {
             flex-direction: column;
@@ -441,6 +436,13 @@
               margin-left: 0;
             }
           }
+        }
+
+        .buttons {
+          display: flex;
+          flex-direction: column;
+          position: relative;
+          gap: 30px;
         }
       }
     }
